devtools/client/shared/components/reps/test/mochitest/test_reps_promise.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Promise rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Promise</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
"use strict";

window.onload = Task.async(function* () {
  const {
    REPS,
    MODE,
    getRep,
    getSelectableInInspectorGrips,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { PromiseRep } = REPS;

  const componentUnderTest = PromiseRep;

  try {
    yield testPending();
    yield testFulfilledWithNumber();
    yield testFulfilledWithString();
    yield testFulfilledWithObject();
    yield testFulfilledWithArray();

    yield testOnDomNodeMouseOver();
    yield testOnDomNodeMouseOut();
    yield testOnDomNodeInspectIconClick();

    yield testObjectLink();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testPending() {
    // Test object = `new Promise((resolve, reject) => true)`
    const stub = getGripStub("testPending");

    // Test that correct rep is chosen.
    is(getRep(stub), PromiseRep.rep,
      "Rep correctly selects PromiseRep Rep for pending Promise");

    // Test rendering
    const defaultOutput = `Promise { <state>: "pending" }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `Promise { "pending" }`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, "testPending", componentUnderTest, stub);
  }
  function testFulfilledWithNumber() {
    // Test object = `Promise.resolve(42)`
    const stub = getGripStub("testFulfilledWithNumber");

    // Test that correct rep is chosen.
    is(getRep(stub), PromiseRep.rep,
      "Rep correctly selects PromiseRep Rep for Promise fulfilled with a number");

    // Test rendering
    const defaultOutput = `Promise { <state>: "fulfilled", <value>: 42 }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `Promise { "fulfilled" }`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, "testFulfilledWithNumber", componentUnderTest, stub);
  }
  function testFulfilledWithString() {
    // Test object = `Promise.resolve("foo")`
    const stub = getGripStub("testFulfilledWithString");

    // Test that correct rep is chosen.
    is(getRep(stub), PromiseRep.rep,
      "Rep correctly selects PromiseRep Rep for Promise fulfilled with a string");

    // Test rendering
    const defaultOutput = `Promise { <state>: "fulfilled", <value>: "foo" }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `Promise { "fulfilled" }`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, "testFulfilledWithString", componentUnderTest, stub);
  }

  function testFulfilledWithObject() {
    // Test object = `Promise.resolve({foo: "bar", baz: "boo"})`
    const stub = getGripStub("testFulfilledWithObject");

    // Test that correct rep is chosen.
    is(getRep(stub), PromiseRep.rep,
      "Rep correctly selects PromiseRep Rep for Promise fulfilled with an object");

    // Test rendering
    const defaultOutput = `Promise { <state>: "fulfilled", <value>: Object }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `Promise { "fulfilled" }`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, "testFulfilledWithObject", componentUnderTest, stub);
  }

  function testFulfilledWithArray() {
    // Test object = `Promise.resolve([1,2,3])`
    const stub = getGripStub("testFulfilledWithArray");

    // Test that correct rep is chosen.
    is(getRep(stub), PromiseRep.rep,
      "Rep correctly selects PromiseRep Rep for Promise fulfilled with an array");

    // Test rendering
    const defaultOutput = `Promise { <state>: "fulfilled", <value>: [3] }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `Promise { "fulfilled" }`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, "testFulfilledWithArray", componentUnderTest, stub);
  }

  function testOnDomNodeMouseOver() {
    const stub = getGripStub("testFulfilledWithNode");

    const grips = getSelectableInInspectorGrips(stub);
    is(grips.length, 1, "the stub has one node grip");

    let mouseOverValue;
    let onDOMNodeMouseOver = (object) => {
      mouseOverValue = object;
    };

    const renderedComponent = renderComponent(PromiseRep.rep, {
      object: stub,
      onDOMNodeMouseOver,
    });

    const node = renderedComponent.querySelector(".objectBox-node");
    TestUtils.Simulate.mouseOver(node);

    is(mouseOverValue, grips[0], "onDOMNodeMouseOver is called with " +
      "the expected argument when mouseover is fired on the node element");
  }

  function testOnDomNodeMouseOut() {
    const stub = getGripStub("testFulfilledWithNode");

    const grips = getSelectableInInspectorGrips(stub);
    is(grips.length, 1, "the stub has one node grip");

    let called = false;
    let onDOMNodeMouseOut = (object) => {
      called = true;
    };
    const renderedComponent = renderComponent(PromiseRep.rep, {
      object: stub,
      onDOMNodeMouseOut,
    });

    const node = renderedComponent.querySelector(".objectBox-node");
    TestUtils.Simulate.mouseOut(node);

    is(called, true,
      "onDOMNodeMouseOut is called when mouseout is fired on the node element");
  }

  function testOnDomNodeInspectIconClick() {
    const stub = getGripStub("testFulfilledWithNode");
    const grips = getSelectableInInspectorGrips(stub);
    is(grips.length, 1, "the stub has one node grip");

    let inspectIconClickedValues = null;
    let onInspectIconClick = (object) => {
      inspectIconClickedValues = object;
    };

    let renderedComponentWithoutInspectIcon = renderComponent(PromiseRep.rep, {
      object: getGripStub("testFulfilledWithDisconnectedNode"),
      onInspectIconClick,
    });
    is(renderedComponentWithoutInspectIcon.querySelector(".open-inspector"), null,
      "There isn't an inspect icon when the node is not connected to the DOM tree");

    const renderedComponent = renderComponent(PromiseRep.rep, {
      object: stub,
      onInspectIconClick,
    });

    const icon = renderedComponent.querySelector(".open-inspector");
    ok(icon !== null, "There is an inspect icon as expected");

    TestUtils.Simulate.click(icon);
    is(inspectIconClickedValues, grips[0],
      "onInspectIconClick is called with the expected argument " +
      "when the inspect icon is clicked");
  }

  function testObjectLink() {
    // Test object = `new Promise((resolve, reject) => true)`
    const stub = getGripStub("testPending");

    // Test rendering
    const defaultOutput = `*Promise** { *<state>: "pending"* }*`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `*Promise** { *"pending"* }*`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(
      modeTests,
      "testObjectLink",
      componentUnderTest,
      stub,
      {
        objectLink: (props, ...children) => React.DOM.span({},
          "*", ...children, "*"),
      }
    );
  }

  function getGripStub(name) {
    switch (name) {
      case "testPending":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj54",
          "class": "Promise",
          "promiseState": {
            "state": "pending",
            "creationTimestamp": 1477327760242.5752
          },
          "ownPropertyLength": 0,
          "preview": {
            "kind": "Object",
            "ownProperties": {},
            "ownPropertiesLength": 0,
            "safeGetterValues": {}
          }
        };
      case "testFulfilledWithNumber":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj55",
          "class": "Promise",
          "promiseState": {
            "state": "fulfilled",
            "value": 42,
            "creationTimestamp": 1477327760242.721,
            "timeToSettle": 0.018497000000479602
          },
          "ownPropertyLength": 0,
          "preview": {
            "kind": "Object",
            "ownProperties": {},
            "ownPropertiesLength": 0,
            "safeGetterValues": {}
          }
        };
      case "testFulfilledWithString":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj56",
          "class": "Promise",
          "promiseState": {
            "state": "fulfilled",
            "value": "foo",
            "creationTimestamp": 1477327760243.2483,
            "timeToSettle": 0.0019969999998465937
          },
          "ownPropertyLength": 0,
          "preview": {
            "kind": "Object",
            "ownProperties": {},
            "ownPropertiesLength": 0,
            "safeGetterValues": {}
          }
        };
      case "testFulfilledWithObject":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj59",
          "class": "Promise",
          "promiseState": {
            "state": "fulfilled",
            "value": {
              "type": "object",
              "actor": "server1.conn1.child1/obj60",
              "class": "Object",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 2
            },
            "creationTimestamp": 1477327760243.2214,
            "timeToSettle": 0.002035999999861815
          },
          "ownPropertyLength": 0,
          "preview": {
            "kind": "Object",
            "ownProperties": {},
            "ownPropertiesLength": 0,
            "safeGetterValues": {}
          }
        };
      case "testFulfilledWithArray":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj57",
          "class": "Promise",
          "promiseState": {
            "state": "fulfilled",
            "value": {
              "type": "object",
              "actor": "server1.conn1.child1/obj58",
              "class": "Array",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 4,
              "preview": {
                "kind": "ArrayLike",
                "length": 3
              }
            },
            "creationTimestamp": 1477327760242.9597,
            "timeToSettle": 0.006158000000141328
          },
          "ownPropertyLength": 0,
          "preview": {
            "kind": "Object",
            "ownProperties": {},
            "ownPropertiesLength": 0,
            "safeGetterValues": {}
          }
        };
      case "testFulfilledWithNode":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj217",
          "class": "Promise",
          "promiseState": {
            "state": "fulfilled",
            "value": {
              "type": "object",
              "actor": "server1.conn1.child1/obj218",
              "class": "HTMLButtonElement",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 0,
              "preview": {
                "kind": "DOMNode",
                "nodeType": 1,
                "nodeName": "button",
                "isConnected": true,
                "attributes": {
                  "id": "btn-1",
                  "class": "btn btn-log",
                  "type": "button"
                },
                "attributesLength": 3
              }
            },
            "creationTimestamp": 1480423091620.3716,
            "timeToSettle": 0.02842400000372436
          },
          "ownPropertyLength": 0,
          "preview": {
            "kind": "Object",
            "ownProperties": {},
            "ownPropertiesLength": 0,
            "safeGetterValues": {}
          }
        };
      case "testFulfilledWithDisconnectedNode":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj217",
          "class": "Promise",
          "promiseState": {
            "state": "fulfilled",
            "value": {
              "type": "object",
              "actor": "server1.conn1.child1/obj218",
              "class": "HTMLButtonElement",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 0,
              "preview": {
                "kind": "DOMNode",
                "nodeType": 1,
                "nodeName": "button",
                "isConnected": false,
                "attributes": {
                  "id": "btn-1",
                  "class": "btn btn-log",
                  "type": "button"
                },
                "attributesLength": 3
              }
            },
            "creationTimestamp": 1480423091620.3716,
            "timeToSettle": 0.02842400000372436
          },
          "ownPropertyLength": 0,
          "preview": {
            "kind": "Object",
            "ownProperties": {},
            "ownPropertiesLength": 0,
            "safeGetterValues": {}
          }
        };
    }
    return null;
  }
});
</script>
</pre>
</body>
</html>
